+++ api/helpers/playlists/start.js
@@ -137,7 +137,7 @@ module.exports = {
           await sails.helpers.songs.remove(true, sails.config.custom.subcats.noClearGeneral, true) // Leave requests in the queue for standard playlists.
           await sails.helpers.rest.cmd('EnableAssisted', 0)
           var attendance = await sails.models.attendance.createRecord(`Playlist: ${theplaylist.name}`)
-          await sails.models.meta.changeMeta({ state: 'automation_playlist', playlist: theplaylist.name, playlist_position: -1, playlist_played: moment().toISOString(true) })
+          await sails.models.meta.changeMeta({ state: 'automation_playlist', playlist: theplaylist.name, playlistPosition: -1, playlistPlayed: moment().toISOString(true) })
           await sails.models.logs.create({ attendanceID: sails.models.meta['A'].attendanceID, logtype: 'primary', loglevel: 'success', logsubtype: 'playlist - ' + theplaylist.name, event: '<strong>Playlist started.</strong><br />Playlist: ' + inputs.name }).fetch()
             .tolerate((err) => {
               sails.log.error(err)
@@ -156,7 +156,7 @@ module.exports = {
           await sails.helpers.rest.cmd('EnableAutoDJ', 0)
           await sails.helpers.songs.remove(true, sails.config.custom.subcats.noClearShow, false, false)
           await sails.helpers.rest.cmd('EnableAssisted', 0)
-          await sails.models.meta.changeMeta({ state: 'automation_prerecord', playlist: theplaylist.name, playlist_position: -1, playlist_played: moment().toISOString(true), show: theplaylist.name, topic: await sails.helpers.truncateText(inputs.topic, 256) })
+          await sails.models.meta.changeMeta({ state: 'automation_prerecord', playlist: theplaylist.name, playlistPosition: -1, playlistPlayed: moment().toISOString(true), show: theplaylist.name, topic: await sails.helpers.truncateText(inputs.topic, 256) })
           await loadPlaylist()
 
           // After loading playlist, determine if we should immediately skip the currently playing track to get the prerecord on the air sooner.